python/wifiSwitcher/requirements.txt
psutil==5.9.8
python-dotenv==1.0.1
